move Composite impls into their own file
